requirements.txt
Babel==1.3
dj-database-url==0.3.0
dj-static==0.0.5
Django==1.8.18
django-admin-bootstrapped==2.5.7
django-admin-sortable2==0.2.7
django-bootstrap-form==3.1
django-debug-toolbar==1.5
django-form-admin==0.4.2
django-mailer==0.1.0
django-mathfilters==0.3.0
django-rosetta==0.7.8
django-suit==0.2.7
django-suit-redactor==0.0.2
django-tables2==1.0.6
enum34==1.1.6
google-api-python-client==1.6.4
gunicorn==19.4.5
html2text==2014.7.3
html5lib==0.999
httplib2==0.10.3
jdcal==1.2
mysqlclient==1.3.12
oauth2client==4.1.2
openpyxl==2.2.6
Pillow==2.7.0
polib==1.0.7
psutil==2.1.1
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycparser==2.14
PyDrive==1.0.1
pyPdf==1.13
pystache==0.5.3
python-dateutil==2.2
pytz==2014.2
PyYAML==3.11
reportlab==3.4.0
requests==2.2.1
simplejson==3.8.0
six==1.11.0
sqlparse==0.1.11
static==1.0.2
uritemplate==3.0.0
whitenoise==3.0
xlrd==1.1.0
xlutils==2.0.0
xlwt==1.3.0
